Create a new InsertionService in main.go
