+++ postgresql/connection.go
@@ -2,18 +2,18 @@ package postgresql
 
 import (
 	"database/sql"
-	"errors"
 	"fmt"
-	"os"
 
 	_ "github.com/GoogleCloudPlatform/cloudsql-proxy/proxy/dialers/postgres" // blank import
+
+	"github.com/benchttp/worker"
 )
 
 type InsertionService struct {
 	db *sql.DB
 }
 
-func NewInsertionService(config Config) (InsertionService, error) {
+func NewInsertionService(config worker.Config) (InsertionService, error) {
 	dbURI := fmt.Sprintf("host=%s user=%s password=%s dbname=%s sslmode=disable",
 		config.Host,
 		config.User,
@@ -34,42 +34,4 @@ func NewInsertionService(config Config) (InsertionService, error) {
 	db.SetMaxOpenConns(config.MaxConn)
 
 	return InsertionService{db}, nil
-}
-
-type Config struct {
-	Host     string
-	User     string
-	Password string
-	DBName   string
-	IdleConn int
-	MaxConn  int
-}
-
-func GetConfigFromEnvVariables() (Config, error) {
-	var config Config
-
-	config.Host = os.Getenv("PSQL_HOST")
-	if config.Host == "" {
-		return config, errors.New("PSQL_HOST environment variable not found")
-	}
-
-	config.User = os.Getenv("PSQL_USER")
-	if config.User == "" {
-		return config, errors.New("PSQL_USER environment variable not found")
-	}
-
-	config.Password = os.Getenv("PSQL_PASSWORD")
-	if config.Password == "" {
-		return config, errors.New("PSQL_PASSWORD environment variable not found")
-	}
-
-	config.DBName = os.Getenv("PSQL_NAME")
-	if config.DBName == "" {
-		return config, errors.New("PSQL_NAME environment variable not found")
-	}
-
-	config.IdleConn = 10
-	config.MaxConn = 25
-
-	return config, nil
 }
